[refactor] modal 디렉토리 분리
