--- app/view/modal.php
@@ -1,10 +1,0 @@
-<div id="modal">
-    <div id="modal-background"></div>
-    <div id="modal-content">
-        <?php
-        include_once 'app/view/signup.php';
-        include_once 'app/view/signin.php';
-        ?>
-    </div>
-</div>
-<script src="app/view/modal.js" type="module"></script>
