+++ app/view/modal/modal.js
