+++ app/view/modal/modal.php
@@ -1,0 +1,10 @@
+<div id="modal">
+    <div id="modal-background"></div>
+    <div id="modal-content">
+        <?php
+        include_once 'app/view/modal/signup.php';
+        include_once 'app/view/modal/signin.php';
+        ?>
+    </div>
+</div>
+<script src="app/view/modal/modal.js" type="module"></script>
